more extensive messages, single writer

+++ src/libmrc/include/mrc_redist.h
@@ -34,7 +34,7 @@ struct mrc_redist_writer {
   int writer_rank;
   struct mrc_redist_block *blocks_begin, *blocks_end;
   void *buf;
-  int buf_size;
+  size_t buf_size;
 };
 
 struct mrc_redist_write_send {
